memory on writes
