+++ light/light.cc
@@ -29,7 +29,7 @@
 #include <light/surflight.hh> //mxd
 #include <light/entities.hh>
 #include <light/ltface.hh>
-#include <light/litfile.hh> // for facesup_t
+#include <light/write.hh> // for facesup_t
 #include <light/trace_embree.hh>
 
 #include <common/log.hh>
@@ -65,7 +65,7 @@ static std::span<lightsurf_t> light_surfaces_span;
 // light_surfaces filtered down to just the emissive ones
 static std::vector<lightsurf_t*> emissive_light_surfaces;
 
-std::span<lightsurf_t, std::dynamic_extent> &LightSurfaces()
+std::span<lightsurf_t> &LightSurfaces()
 {
     return light_surfaces_span;
 }
@@ -86,34 +86,13 @@ static void UpdateEmissiveLightSurfacesList()
     }
 }
 
-static std::vector<facesup_t> faces_sup; // lit2/bspx stuff
-static std::vector<bspx_decoupled_lm_perface> facesup_decoupled_global;
+std::vector<facesup_t> faces_sup; // lit2/bspx stuff
+std::vector<bspx_decoupled_lm_perface> facesup_decoupled_global;
 
 bool IsOutputtingSupplementaryData()
 {
     return !faces_sup.empty();
 }
-
-/// start of lightmap data
-std::vector<uint8_t> filebase;
-/// offset of start of free space after data (should be kept a multiple of 4)
-static int file_p;
-/// offset of end of free space for lightmap data
-static int file_end;
-
-/// start of litfile data
-std::vector<uint8_t> lit_filebase;
-/// offset of start of free space after litfile data (should be kept a multiple of 12)
-static int lit_file_p;
-/// offset of end of space for litfile data
-static int lit_file_end;
-
-/// start of luxfile data
-std::vector<uint8_t> lux_filebase;
-/// offset of start of free space after luxfile data (should be kept a multiple of 12)
-static int lux_file_p;
-/// offset of end of space for luxfile data
-static int lux_file_end;
 
 static std::unordered_map<int, std::vector<uint8_t>> all_uncompressed_vis;
 
@@ -554,82 +533,6 @@ void FixupGlobalSettings()
     }
 }
 
-static std::mutex light_mutex;
-
-/*
- * Return space for the lightmap and colourmap at the same time so it can
- * be done in a thread-safe manner.
- *
- * size is the number of greyscale pixels = number of bytes to allocate
- * and return in *lightdata
- */
-void GetFileSpace(uint8_t **lightdata, uint8_t **colordata, uint8_t **deluxdata, int size)
-{
-    light_mutex.lock();
-
-    *lightdata = *colordata = *deluxdata = nullptr;
-
-    if (!filebase.empty()) {
-        *lightdata = filebase.data() + file_p;
-    }
-    if (!lit_filebase.empty()) {
-        *colordata = lit_filebase.data() + lit_file_p;
-    }
-    if (!lux_filebase.empty()) {
-        *deluxdata = lux_filebase.data() + lux_file_p;
-    }
-
-    // if size isn't a multiple of 4, round up to the next multiple of 4
-    if ((size % 4) != 0) {
-        size += (4 - (size % 4));
-    }
-
-    // increment the next writing offsets, aligning them to 4 uint8_t boundaries (file_p)
-    // and 12-uint8_t boundaries (lit_file_p/lux_file_p)
-    if (!filebase.empty()) {
-        file_p += size;
-    }
-    if (!lit_filebase.empty()) {
-        lit_file_p += 3 * size;
-    }
-    if (!lux_filebase.empty()) {
-        lux_file_p += 3 * size;
-    }
-
-    light_mutex.unlock();
-
-    if (file_p > file_end)
-        FError("overrun");
-
-    if (lit_file_p > lit_file_end)
-        FError("overrun");
-}
-
-/**
- * Special version of GetFileSpace for when we're relighting a .bsp and can't modify it.
- * In this case the offsets are already known.
- */
-void GetFileSpace_PreserveOffsetInBsp(uint8_t **lightdata, uint8_t **colordata, uint8_t **deluxdata, int lightofs)
-{
-    Q_assert(lightofs >= 0);
-
-    *lightdata = *colordata = *deluxdata = nullptr;
-
-    if (!filebase.empty()) {
-        *lightdata = filebase.data() + lightofs;
-    }
-
-    if (colordata && !lit_filebase.empty()) {
-        *colordata = lit_filebase.data() + (lightofs * 3);
-    }
-
-    if (deluxdata && !lux_filebase.empty()) {
-        *deluxdata = lux_filebase.data() + (lightofs * 3);
-    }
-
-    // NOTE: file_p et. al. are not updated, since we're not dynamically allocating the lightmaps
-}
-
 const modelinfo_t *ModelInfoForModel(const mbsp_t *bsp, int modelnum)
 {
     return modelinfo.at(modelnum);
@@ -734,44 +637,6 @@ static void CreateLightmapSurfaces(mbsp_t *bsp)
     });
 }
 
-static void SaveLightmapSurfaces(mbsp_t *bsp)
-{
-    logging::funcheader();
-    logging::parallel_for(static_cast<size_t>(0), bsp->dfaces.size(), [&bsp](size_t i) {
-        auto &surf = light_surfaces[i];
-
-        if (surf.samples.empty()) {
-            return;
-        }
-
-        FinishLightmapSurface(bsp, &surf);
-
-        auto f = &bsp->dfaces[i];
-        const modelinfo_t *face_modelinfo = ModelInfoForFace(bsp, i);
-
-        if (!facesup_decoupled_global.empty()) {
-            SaveLightmapSurface(
-                bsp, f, nullptr, &facesup_decoupled_global[i], &surf, surf.extents, surf.extents);
-        } else if (faces_sup.empty()) {
-            SaveLightmapSurface(bsp, f, nullptr, nullptr, &surf, surf.extents, surf.extents);
-        } else if (light_options.novanilla.value() || faces_sup[i].lmscale == face_modelinfo->lightmapscale) {
-            if (faces_sup[i].lmscale == face_modelinfo->lightmapscale) {
-                f->lightofs = faces_sup[i].lightofs;
-            } else {
-                f->lightofs = -1;
-            }
-            SaveLightmapSurface(bsp, f, &faces_sup[i], nullptr, &surf, surf.extents, surf.extents);
-            for (int j = 0; j < MAXLIGHTMAPS; j++) {
-                f->styles[j] =
-                    faces_sup[i].styles[j] == INVALID_LIGHTSTYLE ? INVALID_LIGHTSTYLE_OLD : faces_sup[i].styles[j];
-            }
-        } else {
-            SaveLightmapSurface(bsp, f, nullptr, nullptr, &surf, surf.extents, surf.vanilla_extents);
-            SaveLightmapSurface(bsp, f, &faces_sup[i], nullptr, &surf, surf.extents, surf.extents);
-        }
-    });
-}
-
 static void ClearLightmapSurfaces()
 {
     logging::funcheader();
@@ -854,47 +719,18 @@ static void FindModelInfo(const mbsp_t *bsp)
     Q_assert(modelinfo.size() == bsp->dmodels.size());
 }
 
-// FIXME: in theory can't we calculate the exact amount of
-// storage required? we'd have to expand it by 4 to account for
-// lightstyles though
-static constexpr size_t MAX_MAP_LIGHTING = 0x8000000;
-
 /*
  * =============
  *  LightWorld
  * =============
  */
-static void LightWorld(bspdata_t *bspdata, bool forcedscale)
+static void LightWorld(bspdata_t *bspdata, const fs::path &source, bool forcedscale)
 {
     logging::funcheader();
 
     mbsp_t &bsp = std::get<mbsp_t>(bspdata->bsp);
 
     ClearLightmapSurfaces();
-    filebase.clear();
-    lit_filebase.clear();
-    lux_filebase.clear();
-
-    if (!bsp.loadversion->game->has_rgb_lightmap) {
-        /* greyscale data stored in a separate buffer */
-        filebase.resize(MAX_MAP_LIGHTING);
-        file_p = 0;
-        file_end = MAX_MAP_LIGHTING;
-    }
-
-    if (bsp.loadversion->game->has_rgb_lightmap || light_options.write_litfile) {
-        /* litfile data stored in a separate buffer */
-        lit_filebase.resize(MAX_MAP_LIGHTING * 3);
-        lit_file_p = 0;
-        lit_file_end = (MAX_MAP_LIGHTING * 3);
-    }
-
-    if (light_options.write_luxfile) {
-        /* lux data stored in a separate buffer */
-        lux_filebase.resize(MAX_MAP_LIGHTING * 3);
-        lux_file_p = 0;
-        lux_file_end = (MAX_MAP_LIGHTING * 3);
-    }
 
     if (forcedscale) {
         bspdata->bspx.entries.erase("LMSHIFT");
@@ -996,23 +832,7 @@ static void LightWorld(bspdata_t *bspdata, bool forcedscale)
         });
     }
 
-    SaveLightmapSurfaces(&bsp);
-
-    logging::print("Lighting Completed.\n\n");
-
-    // Transfer greyscale lightmap (or color lightmap for Q2/HL) to the bsp and update lightdatasize
-    if (!light_options.litonly.value()) {
-        if (bsp.loadversion->game->has_rgb_lightmap) {
-            bsp.dlightdata.resize(lit_file_p);
-            memcpy(bsp.dlightdata.data(), lit_filebase.data(), bsp.dlightdata.size());
-        } else {
-            bsp.dlightdata.resize(file_p);
-            memcpy(bsp.dlightdata.data(), filebase.data(), bsp.dlightdata.size());
-        }
-    } else {
-        // NOTE: bsp.lightdatasize is already valid in the -litonly case
-    }
-    logging::print("lightdatasize: {}\n", bsp.dlightdata.size());
+    SaveLightmapSurfaces(bspdata, source);
 
     // kill this stuff if its somehow found.
     bspdata->bspx.entries.erase("LMSTYLE16");
@@ -1476,18 +1296,6 @@ static void ResetLight()
     faces_sup.clear();
     facesup_decoupled_global.clear();
 
-    filebase.clear();
-    file_p = 0;
-    file_end = 0;
-
-    lit_filebase.clear();
-    lit_file_p = 0;
-    lit_file_end = 0;
-
-    lux_filebase.clear();
-    lux_file_p = 0;
-    lux_file_end = 0;
-
     all_uncompressed_vis.clear();
     modelinfo.clear();
     tracelist.clear();
@@ -1629,7 +1437,7 @@ int light_main(int argc, const char **argv)
 
         SetupDirt(light_options);
 
-        LightWorld(&bspdata, light_options.lightmap_scale.is_changed());
+        LightWorld(&bspdata, source, light_options.lightmap_scale.is_changed());
 
         LightGrid(&bspdata);
 
@@ -1642,29 +1450,8 @@ int light_main(int argc, const char **argv)
             WriteNormals(bsp, bspdata);
         }
 
-        /*invalidate any bspx lighting info early*/
-        bspdata.bspx.entries.erase("RGBLIGHTING");
-        bspdata.bspx.entries.erase("LIGHTINGDIR");
-
         if (light_options.write_litfile == lightfile::lit2) {
-            WriteLitFile(&bsp, faces_sup, source, 2);
             return 0; // run away before any files are written
-        }
-
-        /*fixme: add a new per-surface offset+lmscale lump for compat/versitility?*/
-        if (light_options.write_litfile & lightfile::external) {
-            WriteLitFile(&bsp, faces_sup, source, LIT_VERSION);
-        }
-        if (light_options.write_litfile & lightfile::bspx) {
-            lit_filebase.resize(bsp.dlightdata.size() * 3);
-            bspdata.bspx.transfer("RGBLIGHTING", lit_filebase);
-        }
-        if (light_options.write_luxfile & lightfile::external) {
-            WriteLuxFile(&bsp, source, LIT_VERSION);
-        }
-        if (light_options.write_luxfile & lightfile::bspx) {
-            lux_filebase.resize(bsp.dlightdata.size() * 3);
-            bspdata.bspx.transfer("LIGHTINGDIR", lux_filebase);
         }
     }
